Move parametrizations to a separate file
