+++ tests/test_ewl.py
@@ -1,5 +1,3 @@
-from typing import Dict
-
 import numpy as np
 import pytest
 import sympy as sp
@@ -8,8 +6,8 @@ from qiskit.quantum_info.operators import Operator
 from sympy import Matrix
 from sympy.physics.quantum.qubit import Qubit
 
-from ewl import i, pi, sqrt2, number_of_qubits, sympy_to_numpy_matrix, \
-    U_theta_alpha_beta, U_theta_phi_lambda, U_theta_phi, J, U, EWL
+from ewl import i, pi, sqrt2, number_of_qubits, sympy_to_numpy_matrix, J, EWL
+from ewl.parametrizations import U
 
 
 @pytest.mark.parametrize('psi, expected', [
@@ -24,32 +22,6 @@ def test_number_of_qubits(psi, expected: int):
 def test_sympy_to_numpy_matrix():
     values = [[1, 2], [3, 4]]
     assert np.allclose(sympy_to_numpy_matrix(Matrix(values)), np.array(values))
-
-
-@pytest.mark.parametrize('kwargs, expected', [
-    (dict(theta=0, alpha=0, beta=pi / 2), Matrix([[1, 0], [0, 1]])),
-    (dict(theta=pi / 2, alpha=pi / 2, beta=0), Matrix([[i / sqrt2, i / sqrt2], [i / sqrt2, -i / sqrt2]])),
-    (dict(theta=0, alpha=pi / 2), Matrix([[i, 0], [0, -i]])),
-    (dict(theta=0, alpha=0, beta=0), Matrix([[1, 0], [0, 1]])),
-    (dict(theta=pi, alpha=0, beta=0), Matrix([[0, i], [i, 0]])),
-])
-def test_U_theta_alpha(kwargs: Dict[str, complex], expected: Matrix):
-    assert U_theta_alpha_beta(**kwargs) == expected
-
-
-@pytest.mark.parametrize('kwargs, expected', [
-    (dict(theta=0, phi=0, lambda_=0), Matrix([[1, 0], [0, 1]])),  # C
-    (dict(theta=pi, phi=0, lambda_=0), Matrix([[0, -1], [1, 0]])),  # D
-])
-def test_U_theta_phi_lambda(kwargs: Dict[str, complex], expected: Matrix):
-    assert U_theta_phi_lambda(**kwargs) == expected
-
-
-@pytest.mark.parametrize('kwargs, expected', [
-    (dict(theta=0, phi=pi / 2), Matrix([[i, 0], [0, -i]]))
-])
-def test_U_theta_phi(kwargs: Dict[str, complex], expected: Matrix):
-    assert U_theta_phi(**kwargs) == expected
 
 
 def test_J():
